% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomologyParsing.R
\name{Catalog}
\alias{Catalog}
\title{A function for determining nucleotide overlap that links genes.}
\usage{
Catalog(ResultMatrix, Verbose = FALSE)
}
\arguments{
\item{ResultMatrix}{A matrix where with the upper triangle containing matrices of genes linked by nucleotide overlap}
}
\description{
A function for determining nucleotide overlap that links genes.
}
\examples{
Catalog()
}
\keyword{Core}
\keyword{Genome}
\keyword{Homology,}
\keyword{Orthology,}
