% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreAlignment.R
\name{CoreAligner}
\alias{CoreAligner}
\title{A function for aligning core genomes. Requires a list of matrices, indicating the homologs present, by gene index, in each genome (column).}
\usage{
CoreAligner(ListOfSets, PATH, GeneCalls, Verbose = FALSE)
}
\arguments{
\item{ListOfSets}{A list of matrices}

\item{PATH}{A character string identifying a sqlite database}

\item{GeneCalls}{A list of dataframes, one for each genome, with named columns "Start", "Stop", and "Strand"}
}
\description{
A function for aligning core genomes. Requires a list of matrices, indicating the homologs present, by gene index, in each genome (column).
}
\examples{
CoreAligner()
}
\keyword{Core}
\keyword{Genome}
