% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFFParser.R
\name{GFFParser}
\alias{GFFParser}
\title{A function for parsing genomic General Format Files}
\usage{
GFFParser(GFFAddress, Verbose = FALSE)
}
\arguments{
\item{GFFAddress}{a matrix of lists, the upper triange of which is filled with matrices of putative homologs}
}
\description{
A function for parsing genomic General Format Files
}
\examples{
GFFParser()
}
\keyword{GFF}
