% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogicalPan.R
\name{LogicalPan}
\alias{LogicalPan}
\title{A function for parsing genomic General Format Files}
\usage{
LogicalPan(HomologList, GeneCalls, Verbose = FALSE, Plot = FALSE)
}
\arguments{
\item{HomologList}{a list of matrices that represents sets of homologs}

\item{GeneCalls}{a list of matrices that represent gene calls}
}
\description{
A function for parsing genomic General Format Files
}
\examples{
LogicalPan()
}
\keyword{Homologs}
