% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucleotideOverlap.R
\name{NucleotideOverlap}
\alias{NucleotideOverlap}
\title{A function for determining nucleotide overlap that links genes.}
\usage{
NucleotideOverlap(SyntenyObject, GeneCalls, PATH, Verbose = FALSE)
}
\arguments{
\item{SyntenyObject}{An object of class "Synteny"}

\item{GeneCalls}{A list of dataframes, one for each genome, with named columns "Start", "Stop", and "Strand"}

\item{PATH}{A character string identifying a sqlite database}
}
\description{
A function for determining nucleotide overlap that links genes.
}
\examples{
NucleotideOverlap()
}
\keyword{Homology}
