% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomologyParsing.R
\name{Catalog}
\alias{Catalog}
\title{A function for doing simple network analysis on sets of homologs}
\usage{
Catalog(ResultMatrix, Verbose = FALSE)
}
\arguments{
\item{ResultMatrix}{a matrix of lists, the upper triange of which is filled with matrices of putative homologs}
}
\description{
A function for doing simple network analysis on sets of homologs
}
\examples{
Catalog()
}
\keyword{Homology}
